Delete DaemonServer; relocate its remaining concerns to free functions
DaemonServer had been reduced to three unrelated concerns. Distribute each
to its natural home and remove the class entirely:

- start(): inline its provider init + identity sync into the daemon startup
  sequence in lifecycle. The IDENTITY.md read/parse/sync becomes
  syncWorkspaceIdentityToPlatform() in platform/sync-identity.ts, beside the
  on-change sync it mirrors.
- broadcastStatus(): becomes broadcastDaemonStatus() in a small daemon-status
  module that owns the package-version read.
- getConversationForMessages(): was a pass-through; callers now use
  getOrCreateConversation() from conversation-store directly.

No behavior change: startup runs the same provider init and identity sync in
the same order, status is broadcast at the same point, and conversation
lookups resolve through the same store function.
